backend/requirements.txt
# 暗标竞标系统 - 后端依赖

# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 加密
cryptography==41.0.7
pycryptodome==3.19.0

# 工具
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 日志
loguru==0.7.2

# CORS
fastapi-cors==0.0.6
